requirements.txt
blinker==1.9.0
click==8.3.0
Flask==3.1.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.3.3
packaging==25.0
parameterized==0.9.0
pluggy==1.6.0
Pygments==2.19.2
PyMuPDF==1.26.5
pytest==8.4.2
python-dotenv==1.1.1
waitress==3.0.2
Werkzeug==3.1.3
